+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Fixed
 
 * documentation building on readthedocs is fixed
+* fixed issue with newer version of fsspec and put/get
 
 ### Removed